requirements.txt
craftax
jax
networkx
numpy
tqdm
pillow
gym
matplotlib
python-dotenv
yandex_cloud_ml_sdk
langchain
langchain-community
toolz
msgpack
seaborn
chromadb
black
